+++ akshare/__init__.py
@@ -3259,6 +3259,7 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.18.51 fix: fix stock_a_gxl_lg interface
 1.18.52 fix: fix fund_scale_daily_szse interface
 1.18.53 fix: fix futures_hist_daily_cffex interface
+1.18.54 fix: fix stock_individual_notice_report interface
 """
 
 from akshare._version import __version__
@@ -4628,7 +4629,10 @@ from akshare.stock_feature.stock_lh_yybpm import (
 """
 沪深 A 股公告
 """
-from akshare.stock_fundamental.stock_notice import stock_notice_report
+from akshare.stock_fundamental.stock_notice import (
+    stock_notice_report,
+    stock_individual_notice_report,
+)
 
 """
 首发企业申报